\name{calculateThreshold }\alias{calculateThreshold }\title{Calculate PPA significance threshold leading to a desired false discovery rate}\description{In the context of multiple testing and discoveries, a popular approach is to use a common threshold leading to a desired false discovery rate (FDR). In the Bayesian paradigm, derivation of the PPA threshold is trivial and can be calculated using a direct posterior probability calculation as described in Newton et al. (2004).}\usage{calculateFDR(prob, threshold)}\arguments{  \item{prob}{matrix or data frame that contains Posterior Probability of Association  (output of eqtl.mcmc function).}  \item{ threshold)}{ The desired false discovery rate. Ex: 0.1 for 10%}}\value{\item{cutoff }{The significance threshold value }}\examples{#To calculate the threshold leading to 10% FDRPPA= eqtl.mcmc(snp_exp, expr_exp, n.iter=1000,burn.in=1000,n.sweep=10, ncores=2)ThresholdPPA =calculateThreshold(PPA,0.1)}