\name{eqtl.mcmc}\alias{eqtl.mcmc}\title{Bayesian Multiple eQTL mapping using MCMC}\description{Compute MCMC algorithm to produce Posterior Probability of Association for   eQTL mapping.   }\usage{eqtl.mcmc(snp, expr, n.iter, burn.in, n.sweep, nproc,  constC = TRUE, write.output = TRUE)}%- maybe also 'usage' for other objects documented here.\arguments{  \item{snp}{ Data frame or matrix with genotype data}  \item{expr}{ Data frame or matrix phenotype data}  \item{n.iter}{Number of iterations to be saved}  \item{burn.in}{The initial burn-in period}  \item{n.sweep}{The sweeping number which must be a positive number}  \item{nproc}{The number of threads you would like to use for   	parallel processing}   \item{constC}{Should variables C be fixed? If TRUE, fixed to   	number of SNPS. If FALSE, C parameters are given independent  	inverse gamma priors.}  \item{write.output}{Write chain iterations to file? If TRUE,   	output for variables will be written to files created in   	the working directory.}
  \item{RIS}{If TRUE, the genotype need to be 0 and 1. If FALSE the genotype need to be 0,1 and 2.}}\value{A matrix with Posterior Probability of Association. Rows correspond to snpsfrom original snp data object, columns correspond to genes from expr data object.}\examples{## Calculate PPA
data(gene)
data(snp)#PPA= eqtl.mcmc(snp, gene, n.iter=1000,burn.in=1000,n.sweep=10,nproc=2,write.output=FALSE)}