\name{gene}
\alias{gene}
\docType{data}
\title{
Gene expression from whole eye tissue from n = 68 BXD RIS mice.
}
\description{
This dataset comprises the profiles of mRNA abundance in whole eye tissue from n = 68 BXD RIS mice, as measured using Affymetrix M430 2.0 microarrays. To ease calculation and facilitate comparisons, we will use a set of G = 1000 probes
}
\usage{data(gene)}
\format{
  The format is:
Formal class 'ExpressionSet' [package "Biobase"] }

\source{
This example uses data generated by Williams and Lu, as available from the Gene Networkwebsite (genenetwork.com). This dataset comprises the profiles of mRNA abundance in whole eye tissue from n = 68 BXD RIS mice, as measured using Affymetrix M430 2.0 microarrays. To ease calculation and facilitate comparisons, we will use a set of G = 1000 probes
}

\examples{
data(gene)
}
\keyword{datasets}
