\name{genepos}
\alias{genepos}
\docType{data}
\title{
Gene position data frame 
}
\description{
A data frame specifying the genomic locations of each gene/probe needs to be prepared with the following columns: gene name, chromosome number, start location (in base pairs) and the location (in base pairs).
}
\format{
  A data frame with 1000 observations with the following columns: gene name, chromosome number, start location (in base pairs) and the location (in base pairs).
  
}
\source{
This example uses data generated by Williams and Lu, as available from the Gene Networkwebsite (genenetwork.com). This dataset comprises the profiles of mRNA abundance in whole eye tissue from n = 68 BXD RIS mice, as measured using Affymetrix M430 2.0 microarrays. To ease calculation and facilitate comparisons, we will use a set of G = 1000 probes
}
\examples{
data(genepos)
}
\keyword{datasets}
