\name{genotype.liver}
\alias{genotype.liver}
\docType{data}
\title{
A set of 290 SNP from 60 F2 mice.
}
\description{
A set of 290 single nucleotide polymorphic markers (SNPs)  from 60 F2 mice. }
\format{
  The format is:
Formal class 'SnpSet' [package "Biobase"] }

\source{
 This F2 cross data set containing genotypic and phenotypic information for60 mice was obtained from the lab of Alan Attie Lab at the University of Wisconsin-Madison.These data are also available at GEO (accession number GSE3330). Only the 5000 most variableexpression traits out of 45,265 transcripts from the liver were used for the current example.
}
\examples{
data(genotype.liver)
}
\keyword{datasets}
