\name{iBMQ-package}
\alias{iBMQ-package}
\alias{iBMQ}
\docType{package}
\title{
iBMQ : An Integrated Hierarchical Bayesian Model for Multivariate eQTL Mapping
}
\description{
This method is designed to detect expression QTLs (eQTLs) by incorporating genotypic and gene expression data into a single model while 1) specifically coping with the high dimensionality of eQTL data (large number of genes), 2) borrowing strength from all gene expression data for the mapping procedures, and 3) controlling the number of false positives to a desirable level.
}
