\name{phenotype.liver}
\alias{phenotype.liver}
\docType{data}
\title{
Gene expression from liver tissue from n = 60 F2 mice.
}
\description{
Gene expression of 5000 probe from liver tissue from n = 60 F2 mice.
}
\format{
  The format is:
Formal class 'ExpressionSet' [package "Biobase"] 
}

\source{
 This F2 cross data set containing genotypic and phenotypic information for60 mice was obtained from the lab of Alan Attie Lab at the University of Wisconsin-Madison.These data are also available at GEO (accession number GSE3330). Only the 5000 most variableexpression traits out of 45,265 transcripts from the liver were used for the current example.
}

\examples{
data(phenotype.liver)
}
\keyword{datasets}
