\name{plotEqtl}
\alias{plotEqtl}
\title{
Plot of the genome-wide eQTL mapping
}
\description{
The plotEqtl function provides a genome-wide plot visualizing the characteristics of each eQTLin terms gene/snp position.
}
\usage{
plot.eqtl(peak)
}
\arguments{
  \item{peak}{
A data frame from the output of the eqtlClassifier function.
}
}