\name{PPA.liver}
\alias{PPA.liver}
\docType{data}
\title{
A matrix with Posterior Probabilities of Association
}
\description{
The result is a matrix with Posterior Probabilities ofAssociation for each gene (row) and SNP (column). The PPA matrix previously calculated with 100,000 iterations for liver tissue from n = 60 F2 mice dataset
}
\format{
 A data frame with 290 observations on the following 5000 variables.
 }
\examples{
data(PPA.liver)
}
\keyword{datasets}
