\name{probe.liver}
\alias{probe.liver}
\docType{data}
\title{
Gene position data frame
}
\description{
A data frame specifying the genomic locations of each gene/probe needs to be prepared with the following columns: gene name, chromosome number, start location (in base pairs) and the location (in base pairs).
}
\usage{data(probe.liver)}
\format{
  A data frame with 4427 observations with the following columns: gene name, chromosome number, start location (in base pairs) and the location (in base pairs).

  
 }

\source{
 This F2 cross data set containing genotypic and phenotypic information for60 mice was obtained from the lab of Alan Attie Lab at the University of Wisconsin-Madison.These data are also available at GEO (accession number GSE3330). Only the 5000 most variableexpression traits out of 45,265 transcripts from the liver were used for the current example.
}

\examples{
data(probe.liver)

}
\keyword{datasets}
