\name{snp}
\alias{snp}
\docType{data}
\title{
A set of 1700 SNP from 68 BXD RIS mice.
}
\description{
A set of 1700 single nucleotide polymorphic markers (SNPs) from 68 BXD RIS mice.
}
\usage{data(snp)}
\format{
  The format is:
Formal class 'SnpSet' [package "Biobase"] }
\source{
This example uses data generated by Williams and Lu, as available from the Gene Networkwebsite (genenetwork.com). This dataset comprises the profiles of mRNA abundance in wholeeye tissue from n = 68 BXD RIS mice, as measured using Affymetrix M430 2.0 microarrays . To ease calculation and facilitate comparisons, we will use a set of G = 1000 probes and1700 single nucleotide polymorphic markers (SNPs).
}

\examples{
data(snp)
}
\keyword{datasets}
