\name{snppos}
\alias{snppos}
\docType{data}
\title{
SNP position data frame 
}
\description{
A data frame specifying the genomiclocations of each SNP with following columns: SNP name, chromosome number, SNP location (in base pair).
}
\usage{data(snppos)}
\format{
  A data frame with 1700 observations with the following columns: SNP name, chromosome number, SNP location (in base pair).
 }

\source{
This example uses data generated by Williams and Lu, as available from the Gene Network website (genenetwork.com). 
}

\examples{
data(snppos)
}
\keyword{datasets}
