\name{calculateThreshold }\alias{calculateThreshold }\title{Calculate PPA significance threshold leading to a desired false discovery rate}\description{In the context of multiple testing and discoveries, a popular approach is to use a common threshold leading to a desired false discovery rate (FDR). In the Bayesian paradigm, derivation of the PPA threshold is trivial and can be calculated using a direct posteriorprobability calculation as described in Newton et al. (2004).}\usage{calculateFDR(prob, threshold)}\arguments{  \item{prob}{matrix or data frame that contains Posterior   Probability of Association  (output of eqtl.mcmc function).}  \item{ threshold)}{ The desired false discovery rate.   Ex: 0.1 for 10\%}}\value{\item{cutoff }{The significance threshold value }}\examples{#To calculate the threshold leading to 10\% FDR#>PPA= eqtl.mcmc(snp, gene, n.iter=100,burn.in=100,n.sweep=20,nproc=6,write.output=FALSE)
>cutoff=calculateThreshold(PPA, 0.1)}