\name{eqtl.mcmc}\alias{eqtl.mcmc}\title{Bayesian Mutiple  eQTL mapping using MCMC}\description{Compute MCMC algorithm to produce Posterior Probability of Association for   eQTL mapping.   }\usage{eqtl.mcmc(snp, expr, n.iter, burn.in, n.sweep, nproc)}%- maybe also 'usage' for other objects documented here.\arguments{  \item{snp}{ Data frame or matrix with genotype data}  \item{expr}{ Data frame or matrix phenotype data}  \item{n.iter}{Number of iterations to be saved}  \item{burn.in}{ The initial burn-in period}  \item{n.sweep}{ The sweeping number which must be a positive number }\item{nproc}{ The number of threads you would like to use for parallel processing}}\value{A matrix with Posterior Probability of Association. }}\examples{## Calculate PPA#PPA= eqtl.mcmc(snp_exp, expr_exp, n.iter=1000,burn.in=1000,n.sweep=10,nproc=2)}