\name{PPA.liver}
\alias{PPA.liver}
\docType{data}
\title{
A matrix with Posterior Probabilities of Association
}
\description{
The result is a matrix with Posterior Probabilities ofAssociation for each gene (row) and SNP (column). The PPA matrix was previously calculated with 100,000 iterations for liver tissue from n = 60 F2 mice dataset
}

\examples{
data(PPA.liver)
}
\keyword{datasets}
