\name{PPA}
\alias{PPA}
\docType{data}
\title{
A matrix with Posterior Probabilities of Association
}
\description{
The result is a matrix with Posterior Probabilities ofAssociation for each gene (row) and SNP (column). The PPA matrix previously calculated with 100,000 iterations for the whole eye tissue from n = 68 BXD RIS mice dataset
}
\format{
  A data frame with 1700 observations on the following 1000 variables.
  }
\examples{
data(PPA)
}
\keyword{datasets}
