\name{snppos}
\alias{snppos}
\docType{data}
\title{
SNP position data frame 
}
\description{
A data frame specifying the genomiclocations of each SNP with following columns: SNP name, chromosome number, SNP location(in base pair).
}
\usage{data(snppos)}
\format{
  A data frame with 1700 observations on the following 3 variables.
  \describe{
    \item{\code{SNP}}{a character vector}
    \item{\code{Chr}}{a numeric vector}
    \item{\code{Mb}}{a numeric vector}
  }
}

\source{
snp pos
}
\references{
snp pos
}
\examples{
data(snppos)

}
\keyword{datasets}
